# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Iqrah Backend API",
    "description": "REST API for the Iqrah mobile app",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/v1/admin/packs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Lists all packs and their latest versions for admin tooling.",
        "operationId": "list_all_packs",
        "responses": {
          "200": {
            "description": "All packs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackManifestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid/disabled admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_api_key": []
          },
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Registers a new pack.",
        "operationId": "register_pack",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Pack registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterPackResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid/disabled admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_api_key": []
          },
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/v1/admin/packs/{id}/disable": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Disables an existing pack (removes it from public availability).",
        "operationId": "disable_pack",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Pack ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pack disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisablePackResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid/disabled admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Pack not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_api_key": []
          },
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/admin/packs/{id}/publish": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Publishes an existing pack.",
        "operationId": "publish_pack",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Pack ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pack published",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishPackResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid/disabled admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_api_key": []
          },
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/admin/packs/{id}/versions": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Uploads a new version file for an existing pack.",
        "operationId": "add_version",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Pack ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/AddVersionMultipartBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Version uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddVersionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid/disabled admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_api_key": []
          },
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "API error response format",
        "required": [
          "error"
        ],
        "properties": {
          "details": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "example": "[\"name: validation failed\"]"
          },
          "error": {
            "type": "string",
            "example": "Validation error: field is required"
          }
        }
      },
...
      "DisablePackResponse": {
        "type": "object",
        "description": "Response for disabling a pack.",
        "required": [
          "disabled"
        ],
        "properties": {
          "disabled": {
            "type": "boolean"
          }
        }
      },
...
      "PackManifestResponse": {
        "type": "object",
        "description": "Global pack manifest response.",
        "required": [
          "packs"
        ],
        "properties": {
          "packs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PackManifestEntry"
            }
          }
        }
      },
...
      "RegisterPackRequest": {
        "type": "object",
        "description": "Request to register a new pack.",
        "required": [
          "name",
          "description",
          "pack_type"
        ],
        "properties": {
          "description": {
            "type": "string",
            "example": "English translation pack for mobile clients"
          },
          "name": {
            "type": "string",
            "example": "English Translation"
          },
          "pack_type": {
            "$ref": "#/components/schemas/PackType"
          }
        }
      },
      "RegisterPackResponse": {
        "type": "object",
        "description": "Response for pack registration.",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        }
      },
...
    },
    "securitySchemes": {
      "admin_api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key"
      },
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}